selftests: mptcp: lib: skip if not below kernel version
Selftests are supposed to run on any kernels, including the old ones not
supporting all MPTCP features.

A new function is now available to easily detect if a feature is
missing by looking at the kernel version. That's clearly not ideal and
this kind of check should be avoided as soon as possible. But sometimes,
there are no external sign that a feature is available or not, e.g. when
the in-kernel path-manager supports having subflows in IPv4 and IPv6 for
the same MPTCP connections, the uAPI has not been modified but the
internal behaviour did. In such case, the only way to verify if the
feature is present is to run the test but then the validation cannot
determine if there is a failure or if the feature is not available. Then
it looks better to check the kernel version instead of having tests that
can never fail.

This function is going to be used in the following commits. In order to
ease the backport of such future patches, it would be good if this
present commit is backported up to the introduction of MPTCP selftests,
hence the Fixes tag below: this type of check was supposed to be done
from the beginning.

Link: https://github.com/multipath-tcp/mptcp_net-next/issues/368
Fixes: 048d19d444be ("mptcp: add basic kselftest for mptcp")
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>