selftests: mptcp: connect: skip transp tests if not supported
Selftests are supposed to run on any kernels, including the old ones not
supporting all MPTCP features.

One of them is the support of IP(V6)_TRANSPARENT socket option with
MPTCP connections. We can skip these tests if the feature is not
supported. For that, we check that TOS is supported: IP(V6)_TRANSPARENT
has been added after TOS support which came with the required
infrastructure in MPTCP sockopt code. To support TOS, the following
function has been exported (T). Not great but better than checking for a
specific kernel version.

Link: https://github.com/multipath-tcp/mptcp_net-next/issues/368
Fixes: 5fb62e9cd3ad ("selftests: mptcp: add tproxy test case")
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>