gh: tests: use linked tag instead of branch in notif
To be able to save the cache, we need to use branches instead of tags.
But for the notif, it might be better to use the tag
(export/20240308T121027), not just the branch (export).

While at it, use the full check of the ref (just in case 'export' or
'export-net' tags are used), and put the URL at the end on IRC (not nice
output when the URL got a preview).

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>