gh: tests: avoid coverage errors
If the test finishes before the end with an error, there is no
coverage.txt file, nor any kernel.lcov. No need to propagate errors to
'Publish conclusion' and 'Coveralls', we can print a message instead,
and create an empty file.

While at it, do not use errexit for this publish job, not to fail for
that.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>