mptcp: pm: rename 'add_addr_signal_max' to 'endp_signal_max'
A few variables linked to the Path-Manager are confusing, and it would
help current and future developers, to clarify them.

One of them is 'add_addr_signal_max', which in fact represents the
maximum number of 'signal' endpoints that can be used to announced
addresses, and not the number of ADD_ADDR that can be signalled.

While at it, add an additional name for the corresponding variable in
MPTCP INFO: mptcpi_endp_signal_max. Not to break the current uAPI, the
new name is added as a 'define' pointing to the former name. This will
then also help userspace devs.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>