mptcp: pm: rename 'add_addr_accept_max' to 'limit_add_addr_accepted'
A few variables linked to the Path-Manager are confusing, and it would
help current and future developers, to clarify them.

One of them is 'add_addr_accept_max', which in fact represents the limit
of ADD_ADDR that can be accepted:  the limit set via 'ip mptcp limit
add_addr_accepted X' for example. It is not linked to the maximum number
of accepted ADD_ADDR.

While at it, add an additional name for the corresponding variable in
MPTCP INFO: mptcpi_limit_add_addr_accepted. Not to break the current
uAPI, the new name is added as a 'define' pointing to the former name.
This will then also help userspace devs.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>