mptcp: pm: in-kernel: reduce pernet struct size
All the 'unsigned int' variables from the 'pm_nl_pernet' structure are
bounded to MPTCP_PM_ADDR_MAX. The endpoint ID is also bounded by the
protocol to 8-bit.

So no need to use 'unsigned int' variables, 'u8' is enough.

Note that the exposed counters in MPTCP_INFO are already limited to
8-bit, same for pm->extra_subflows, and others.

Signed-off-by: Matthieu Baerts (NGI0) <matttbe@kernel.org>